Fix fp16/bf16 dtype crash and vectorize SVD primitives

+++ glassbox/attention_tracker.py
@@ -111,9 +111,11 @@ def compute_attention_tracker_features_matrix_free(
         return apply_AT_blocked(Q, K, u, scale, block_size, causal=causal)
 
     if method == "lanczos":
-        _, S, _ = svd_via_lanczos(matvec, matvec_t, L, k, max(2 * k + 2, 20), str(device))
+        _, S, _ = svd_via_lanczos(
+            matvec, matvec_t, L, k, max(2 * k + 2, 20), str(device), dtype=Q.dtype
+        )
     else:
-        _, S, _ = randomized_svd(matvec, matvec_t, L, k, device=str(device))
+        _, S, _ = randomized_svd(matvec, matvec_t, L, k, device=str(device), dtype=Q.dtype)
 
     S_sorted, _ = torch.sort(S, descending=True)
     sigma2 = S_sorted[1].item() if len(S_sorted) > 1 else 0.0
